+++ src/main/groovy/com/bmuschko/gradle/docker/response/container/InspectContainerResponseHandler.groovy
@@ -19,5 +19,6 @@ class InspectContainerResponseHandler implements ResponseHandler<String, Object>
         String exposedPorts = container.config.exposedPorts ? container.config.exposedPorts.toString() : '[]'
         logger.quiet "ExposedPorts : $exposedPorts"
         logger.quiet "LogConfig : $container.hostConfig.logConfig.type.type"
+        logger.quiet "RestartPolicy : $container.hostConfig.restartPolicy"
     }
 }